+++ features/config/offline/display.feature
@@ -1,7 +1,7 @@
 Feature: display the current offline status
 
   Scenario: default value
-    When I run "git-town offline"
+    When I run "git-town config offline"
     Then it prints:
       """
       false
@@ -9,7 +9,7 @@ Feature: display the current offline status
 
   Scenario: enabled
     Given setting "offline" is "true"
-    When I run "git-town offline"
+    When I run "git-town config offline"
     Then it prints:
       """
       true
@@ -17,7 +17,7 @@ Feature: display the current offline status
 
   Scenario: disabled
     Given setting "offline" is "false"
-    When I run "git-town offline"
+    When I run "git-town config offline"
     Then it prints:
       """
       false
@@ -25,7 +25,7 @@ Feature: display the current offline status
 
   Scenario: invalid value
     Given setting "offline" is "zonk"
-    When I run "git-town offline"
+    When I run "git-town config offline"
     Then it prints:
       """
       Invalid value for git-town.offline: "zonk". Please provide either true or false. Considering false for now.
